tutorial-ma: Update date columns range
The data series in who_covid_19_sit_rep_time_series.csv (used as
the data source for this tutorial), have been expanded with an
extra column. Because of this, the time columns now start on
column with index 4. That is, there are four columns instead of
three before the date columns.

Hence, update the range of date columns so that instead of using
column index 3 to 17, we now use column index 4 to 18.