tutorial-ma: Ignore row with totals
The last change to address in the updated data source is to
compensate for the addition of a new row for the whole country
of China with the total sums of cases per date.

If we don't remove this row from the data before doing the sum,
we will be summing twice, as the filtering done with:

    nbcases[locations[:, 1] == "China"]

will include this row.