Explicitly pin sphinx in conda env (#136)
At time of merging, pip pins sphinx to 4.5 but conda doesn't, which results
in errors when building the site. Suspect that one or more of the packages on
conda-forge has different dependencies than those defined in the pyproject.toml.

Should be reverted in the future when the conda-forge feedstocks result in the
correct environment.